# Utilize RWMutex to speed up operations for LRUExpireCache