**What type of PR is this?**
/kind cleanup

**What this PR does / why we need it**:
Currently LRUExpireCache uses Mutex for both read and write operations.

With Mutex:
BenchmarkSimpleCache-12     	  500000	      3061 ns/op
BenchmarkStripedCache-12    	 1000000	      1457 ns/op

With RWMutex:
BenchmarkSimpleCache-12     	  500000	      2079 ns/op
BenchmarkStripedCache-12    	 5000000	       314 ns/op

```release-note
NONE
```

**Additional documentation e.g., KEPs (Kubernetes Enhancement Proposals), usage docs, etc.**:

```docs

```