@cheftako 2019-09-10
Hi Ted,
  I'm assuming when you say BenchmarkSimpleCache you are talking https://github.com/kubernetes/apiserver/blob/master/pkg/authentication/token/cache/cache_test.go#L35 ?
That is an excellent test for showing a potential improvement, however that test is a somewhat artificial usage model for generating and testing read contention. 
The important question is there a circumstance which would cause Kubernetes to behave the same way as that test?