@tedyu 2019-09-10
@cheftako 
I had similar question on the applicability of BenchmarkSimpleCache.

I wonder if this benchmark can be improved to reflect actual usage in prod.
If it is not feasible to make such change, we should remove this benchmark because it would confuse developers.