@tedyu 2019-09-11
@cheftako 
How about adding a comment in BenchmarkSimpleCache (and other benchmarks in the same file) saying that the performance profile may not match that in the production.

What do you think ?