@cheftako 2019-09-11
> Walter Fender
> How about adding a comment in BenchmarkSimpleCache (and other benchmarks in the same file) saying that the performance profile may not match that in the production.
> 
> What do you think ?

Sounds like a good plan.
Might even be good to suggest running with and without concurrency to better understand the the impact of a change.
Also suggest that this is a tool to test and measure high concurrency of the cache in isolation and not to the Kubernetes usage of the Cache.