@tedyu 2019-09-13
/test pull-kubernetes-integration